backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.10.0
httpx>=0.27.0
pdfplumber>=0.11.0
python-docx>=1.1.2
sqlparse>=0.5.0

# PostgreSQL
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Redis
redis[asyncio]>=5.0.0

# Rate limiting
slowapi>=0.1.9

# Environment variables
python-dotenv>=1.0.0
